Rename old integrations paths
