+++ src/views/product-integration-old/components/categorized-variable-group-list/index.tsx
