+++ src/views/product-integration-old/components/component-tab-content/index.tsx
