+++ src/views/product-integration-old/components/searchable-variable-group-list/index.tsx
