+++ src/views/product-integration-old/components/searchable-variable-group-list/style.module.css
