+++ src/views/product-integration-old/components/variable-group-list/index.tsx
