+++ src/views/product-integration-old/components/variable-group-list/style.module.css
